+++ src/vec3_utils.c
@@ -6,29 +6,29 @@
 /*   By: avast <avast@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/19 12:48:17 by avast             #+#    #+#             */
-/*   Updated: 2023/05/02 11:41:15 by avast            ###   ########.fr       */
+/*   Updated: 2023/05/03 16:27:52 by avast            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/params.h"
 #include "../includes/proto.h" 
-#include "../libft/libft.h"
+#include "../libft/includes/libft.h"
 
-double	vec3_dot(t_vec3 a, t_vec3 b)
+float	vec3_dot(t_vec3 a, t_vec3 b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
 
-/* double	vec3_distance(t_vec3 a, t_vec3 b)
+/* float	vec3_distance(t_vec3 a, t_vec3 b)
 {
 	return (sqrt(pow(b.x - a.x, 2) + pow(b.y - a.y, 2) + pow(b.z - a.z, 2)))
 } */
 
 t_vec3	vec3_normalize(t_vec3 a)
 {
-	double	length;
+	float	length;
 
-	length = sqrt(vec3_dot(a, a));
+	length = sqrtf(vec3_dot(a, a));
 	return (a.xyz / length);
 }
